Rework troubleshooting section, add basic explainer
This is an attempt to make the troubleshooting doc a bit better
structured. It's a bit lacking in concrete details for some of the
common problems (because I forget exactly how they mainfest), but I hope
that we can beef this up as things come up.

I also added a short explainer doc which explains some of the key
concepts.