Rework troubleshooting section, add basic explainer (#2517)
* Rework troubleshooting section, add basic explainer

This is an attempt to make the troubleshooting doc a bit better
structured. It's a bit lacking in concrete details for some of the
common problems (because I forget exactly how they mainfest), but I hope
that we can beef this up as things come up.

I also added a short explainer doc which explains some of the key
concepts.

* Update docs/troubleshooting.md

Co-authored-by: Pepe Iborra <pepeiborra@me.com>

* Reorganize a bit

Co-authored-by: Anton Latukha <anton.latukha@gmail.com>
Co-authored-by: Pepe Iborra <pepeiborra@me.com>
Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>