Add new Ghcide Argument to track Project Root
This commit provides an alternate way to grab the project root/current working directory.

Prior to this commit the relative filepath "." was hard-coded for both
Db and Custom Commands. This results in inconsistent behaviour with how
HLS derives it's hiedb location.

This new argument to the internal Ghcide Arguments, maps from the
executable Arguments `argsCwd` or by grabbing the current working
directory. If the user provides an option to `--cwd` we need to make
sure we make that filepath absolute.

Finally, inside the command handler, if necessary, we will grab the
current working directory. We cannot provide a suitable default for this
argument, therefore we leave it as a `Maybe FilePath`, even though this
path should never be taken.