GHC-9.0 support for hls-tactics-plugin (#2581)
* Initial partially broken ghc9 support for tactics

It compiles and most tests succeed, but some fail.
In particular, the ones where it should suggest `show` fails to find
that as a possible solution and fails to find evidence for `Show a`.

* Enable tactics plugin for nix as well

* Wingman does support ghc9 now

* Fix stack support for tactics ghc-9.0.1

* Enable tests for tactics on ghc-9 on ci

* Actually enable tactics for ghc-9 on nix

* Wingman: Improve test failure messages

Previously when wingman fails to find a solution, the test failure
would say "Timed out when waiting for a message".
Now it instead prints the error message from wingman.

* Get theta

* Make wrapper theta discovery more reliable

* Fix AutoThetaRankN

* Fix FmapJoin and FmapJoinInLet

* Fix MetaBegin

* Cleanup

* Fix merge

More fixing the merge

No, seriously fix the merge

Fix a broken merge

* Need a comma

* Try a better ConPatIn

* Oops

* its a nightmare

* i hate ci

* ok that fixes the conpat

* Maybe this is the end of it

* ci

* refinery

* undo some changes

* no more tctypes?

* maybe it builds now

* omg

* om F g

* expect fail on 9.2

* fix

* fix again

Co-authored-by: Andreas Källberg <anka.213@gmail.com>
Co-authored-by: Javier Neira <atreyu.bbb@gmail.com>