Add an option to run Fourmolu via the CLI interface of a separate binary, rather than the bundled library
This makes it easier to set up an environment (e.g. a Nix shell) which uses the same version of Fourmolu in HLS as from the command line.

It can also be useful when Fourmolu has updated and we don't want to wait a month for an HLS release, or re-compile.