Send begin progress message synchronously
Currently the Begin progress message is sent asynchronously, so it can
happen that it's never sent if the async is cancelled immediately
because a new kick has started. This causes trouble in tests and
benchmarks which make assumptions about progress updates.

To address this, we send the Begin progress message synchronously,
and only then do the rest of the progress reporting stuff
(including waiting for the response) asynchronously