Refactor plugin: Prefer code action
`isPreferred` can influence the client side order of code actions. The
idea is that an unused import is likely to be removed and less likely
the warning will be disabled. Therefore actions to remove a single or
all redundant imports should be preferred, so that the client can
prioritize them higher.

Followup of
<https://github.com/haskell/haskell-language-server/pull/3018>