Support optional plugins (#3193)
* Support initial generic config for plugins

This enables opt-in  plugins, i.e. plugins that are not enabled by default

* Support initial generic config for plugins

This enables opt-in  plugins, i.e. plugins that are not enabled by default

* Address review comments and remove optionality in configInitialGenericConfig

I can't remember why this was optional in the first place, which probably means no good reason for it to be

* use defConfig instead of def

* test

* descriptor is the golden source for plugin default configs

Read plugin defaults from plugin descriptor instead of copying into initial config

This is less fragile as now the initial generic config specified in the plugin
descriptor is enforced by construction.

The downside is that now the descriptor is needed to parse the config, which
means a few places where a plugin id was enough now require a plugin descriptor,
which is much harder to get. We paper over this by moving these call sites to
the Action monad.

* redundant import

* simplify

* Build fixes

* Build fixes

* pedantic fixes