Add a mergify action to update PRs (#3418)
After a bit of digging I've worked out what's up with mergify. It's
still working fine, but it's not updating PRs very often now. The reason
for that is that it updates PRs that are in the queue, but to get
_into_ the queue you have to pass CI, which is hard ATM because of our
flaky CI. So we can add a separate update rule that just updates the
labelled PRs, so that they are at least kept up-to-date with master.