fix: remove the `Indexing` progress message when exeption in withHieDb
If any exception was happening in `withHieDb`, the `post` action was
never called, leading to a stale `Indexing` message in the client.

By using `bracket`, we ensure that the `post` message is called in the
event of an exception, cleaning the confusing message in the client.

It does not change the well behaving of the indexing (i.e. if it was
crashing, it is still crashing), but at least user won't be confused by
a stale `Indexing` message.