Simplify selection of GHCs to build on (#3633)
* Simplify selection of GHCs to build on

The haskell setup action supports using the major version, and then it
will select the latest minor version. That should be fine for us - we
don't have to change anything to get new minor versions, and it's not
too much of a problem to get new minor versions implicitly, since they
should be very unlikely to break things.

* update exclusions

* Fixup

* More

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>