Set the root for tests to the test directory
On startup, HLS performs a sanity check to validate the GHC version for
the project is the same as the GHC version used to build HLS.
Due to some oversights, the tests always use HLS's cabal package to
perform this check.
This is (almost) valid for the tests but adds overhead for finding the
GHC version for each integration test in HLS.
Improves the startup time for each integration tests that depends on
hls-test-utils. Thus, ghcide test suite is unaffected.

Fix `findCradle` invocations to point to a file in the root directory.
`findCradle` from hie-bios doesn't work correctly for directories and
requires a file, even if that file doesn't exist.