Cleanup func-test suite (#3828)
* Cleanup func-test suite

This test-suite is vastly legacy from Haskell IDE Engine.
It was originally written before we had a plugin-based architecture
where each plugin tests its feature in isolation.
Over time, this test-suite bitrotted, where a lot of testcases were
either redundant or out-of-date and consequentially disabled.

We clean up the test-suite, delete old tests and remove unused testdata.

* Remove unused config change

* Move qualified import tests to hls-refactor-plugin

* Add README for `func-test`

Provides guidance when to add a test to `func-test` and some historical
context.

* Add fendor as CODEOWNER for `func-test` and `hls-test-utils`

---------

Co-authored-by: Michael Peyton Jones <me@michaelpj.com>