Make sure running tests locally pick up the correct cradle type (#3869)
* Run golden tests for "add argument" in temporary directory

If developers have a local `hie.yaml` for HLS development, this causes
the tests to pick up said `hie.yaml`. This causes these tests to use a
cabal cradle, slowing down the test execution. Should have no effect on
CI, though, which never has a `hie.yaml` in the root of the project.

* Remove redundant CPP statement

---------

Co-authored-by: mergify[bot] <37929162+mergify[bot]@users.noreply.github.com>