llvm@11 -> llvm@13
We use `clang` from brew's llvm, and that just fails horribly on Sonoma. 13 seems to work, though we might want to consider to just go with `llvm`. Instead of some ancient pinned version.

With GHC 9.2+ we have the NCG, and don't even need to rely on llvm at all anymore, we could just drop this outright. And use the Apple CLT provided $CC and others.