ormolu-plugin: don’t request GhcSession for useCLI
Probably not the best solution.

When ormolu is requested from the command line, it should do its own
detection of cabal features. If it has to set up the GhcSession just
to get some flags, it blocks saving a file (`formatOnSave = true`),
which is a horrible user experience in cases where the session takes a
minute or more to set up.

Instead, now the plugin only depends on ormolu and the file itself.

Fixes https://github.com/haskell/haskell-language-server/issues/4168