Nix Flake: Remove `hlint`
`hlint` should not be necessary. However, this is an opinion up for debate:

We do not really need the `hlint` executable because it is integrated in HLS.

However, `hlint` is used by a GitHub Action, which we may want to replicate locally.

Also, the `fmt.sh` script is downloading an (unknown) `hlint` version which will
not work on NixOS anyways. In my opinion, if `fmt.sh` is still used (I can not
find any usages) we should fix the script so it is reproducible.