feat: introduce import suggestion for coerce
In a context where we want to `coerce` to a type for which constructor
is not in scope, GHC fails with (e.g., for `Sum Int`):

```
• Couldn't match representation of type ‘Int’ with that of ‘Sum Int’
    arising from a use of ‘coerce’
    The data constructor ‘base-4.18.2.1:Data.Semigroup.Internal.Sum’
      of newtype ‘Sum’ is not in scope
```

This code action detects the missing `newtype` and suggests to add the
required import.

This is convenient because otherwise the user need to interpret the
error message and most of the time manually find which module and type to import.

Note that a better implementation could try to decet that the type is
already imported (if that's the case) and just suggest to add the
constructor (e.g. `(..)`) in the import list, but this is too much
complexity to implement. It could lead to duplicated import lines which
will be "cleaned" by formatter or other extensions.