feat: add test and import suggestion for not in scope record field
In ghc 9.6 we had:

```
Not in scope ‘modelStatusTag’
```

In 9.10 we have:

```
Not in scope: record field ‘modelStatusTag’
```

Introducing the new regex in order to match it AND a test to ensure no
future regression.

The regression is due to the fact that there is a matcher which catch
`Nat in scope: .*`, hence it was matching "record" instead of
"modelStatusTag".