Make the test name shorter
The very long test name leads to overflows in the tasty test output
reporter, causing duplicated lines and generally harder to read output,
when the test output report is displayed.

We refactor the test specification to optionally accept 'TestName's
which can be shorter than the previous way of generating a test name.